+++ src/CommandRunner.ts
@@ -27,7 +27,7 @@ export class CommandRunner extends BaseCommandRunner {
 			if(!TaskUtil.isTaskActive(task, filetype)){
 				return;
 			}
-			xycodeui.channelShow(task.label);
+			xycodeui.info(task.label);
 			let command = await this.commandBuilder.parser(task, configVars, this.file, this.workspaceFolder);
 			let cwd = task.cwd ? this.commandBuilder.format(task.cwd) : this.workspaceFolder;
 			cwd = path.resolve(cwd);
@@ -36,15 +36,22 @@ export class CommandRunner extends BaseCommandRunner {
 				return;
 			}
 			xycodeui.openChannel();
-			xycodeui.channelShow(command);
-			xycodeui.debug(cwd);
-			xycodeui.debug(JSON.stringify(task));
+			xycodeui.info("cwd : " + cwd);
+			xycodeui.debug(task, true);
 			if (task.beforeTriggers) {
 				await this.invokeTrigger(task.beforeTriggers, cwd);
 			}
+
+			xycodeui.debug(Util.isDockerMode ? "docker mode" : Util.isWslMode ? "wsl mode": Util.isBashMode ? "bash mode" : "");
+
 			if (task.termial) {
-				const shellPath = task.termial.shellPath || this.options.shellPath;
-				xycodeui.getTerminal(this.commandBuilder.format(task.termial.name), shellPath, task.termial.shellArgs).sendText(command);
+				let options :any = { 
+					name : this.commandBuilder.format(task.termial.name), 
+					shellPath : task.termial.shellPath || this.options.shellPath,
+					shellArgs : task.termial.shellArgs
+				};
+				xycodeui.debug(options, true);
+				xycodeui.getTerminal(options).sendText(command);
 			}
 			else {
 				let options: child_process.ExecOptions = {
@@ -55,15 +62,20 @@ export class CommandRunner extends BaseCommandRunner {
 				if(task.options){
 					options = { ...options, ...task.options };
 				}
+				if(Util.isWslMode){
+					options.shell = undefined;
+				}
+				xycodeui.debug(options, true);
 				let { stdout, stderr } = await this.exec(command, options);
-				if (stderr && stderr.length > 0) {
-					xycodeui.showErrorMessage(stderr);
-					xycodeui.channelShow(stderr);
+				// It seems some error not show...
+				if (stderr && stderr.length > 0 && stdout !== stderr) {
+					// xycodeui.showErrorMessage(stderr);
+					xycodeui.warn(stderr);
 				}
 				if (stdout && stdout.length > 0) {
 					xycodeui.channelShow(stdout);
 				}
-				if (!(stderr && stderr.length > 0) && task.afterTriggers) {
+				if (!(stderr && stderr.length > 0 && stdout !== stderr) && task.afterTriggers) {
 					await this.invokeTrigger(task.afterTriggers, cwd);
 				}
 			}
@@ -71,7 +83,7 @@ export class CommandRunner extends BaseCommandRunner {
 		}
 		catch (e) {
 			xycodeui.showErrorMessage(`${ExtConst.extName} ${e}`);
-			xycodeui.channelShow(e);
+			xycodeui.fatal(e);
 		}
 	}
 	private async invokeTrigger(triggers: Array<{
@@ -103,20 +115,22 @@ export class CommandBuilder{
 	private configVars: any;
 	private fileAttr: any;
 
-	constructor(private readonly options : CommandRunnerOptions){
+	constructor(private options : CommandRunnerOptions){
 	}
 
 	public async parser(task: TaskType, configVars: any, file: string, workspaceFolder: string) : Promise<string> {
+		XycodeUI.instance.debug("start to parser command");
 		const xycodeui = XycodeUI.instance;
-		this.configVars = { 
-							...configVars, 
-							"SFDX_ALIAS": await new DefaultEnv(this.options).getSfdxAlias()
-						};
-		this.fileAttr = this.getFileAttr(file, workspaceFolder);
+		this.init(file, workspaceFolder);
+		const defaultEnv = await new DefaultEnv(this.options).load(task.command);
+		if(defaultEnv){
+			xycodeui.debug(defaultEnv, true);
+			this.configVars = { ...configVars, ...defaultEnv };
+		}
 
 		let command = task.command;
 		let output = command;
-		const regex = new RegExp(/\${(input|select|multiselect|openFolderDailog|singleFileDailog|multiFilesDailog)(\s)*:(\s)*([^} ]+)(\s)*}/g);
+		const regex = new RegExp(/\${(input|select|multiselect|openFolderDailog|singleFileDailog|multiFilesDailog)(\s)*:(\s)*([^} \|]+)(\s)*(\|wslpath)?}/g);
 		let matches;
 		while ((matches = regex.exec(command)) !== null) {
 			const varkey = matches[4];
@@ -126,8 +140,10 @@ export class CommandBuilder{
 				continue;
 			}
 			const configVar = this.configVars[varkey];
+			const isDockerMode = Util.isDockerMode;
 			let customInput;
 			let customInputWsl;
+			let customInputDocker;
 			if(commandType === "select"){
 				if(!this.configVars.hasOwnProperty(varkey)){
 					throw new Error(`Select Config ${varkey} Error!`);
@@ -143,62 +159,139 @@ export class CommandBuilder{
 			} else if(commandType === "openFolderDailog"){
 				customInput = await xycodeui.openFolderDialog(configVar);
 				customInputWsl = customInput ? Util.getWSLPath(customInput) : undefined;
+				customInputDocker = isDockerMode && customInput ? Util.getDockerPath(customInput) : undefined;
 			} else if(commandType === "singleFileDailog"){
 				customInput = await xycodeui.openFileDialog(configVar);
 				customInputWsl = customInput ? Util.getWSLPath(customInput) : undefined;
+				customInputDocker = isDockerMode && customInput ? Util.getDockerPath(customInput) : undefined;
 			} else if(commandType === "multiFilesDailog"){
-				customInput = await xycodeui.openFilesDialog(configVar, this.options.isWslMode);
-				customInputWsl = customInput;
+				let _openfiles = await xycodeui.openFilesDialog(configVar);
+				let separator = configVar && configVar.hasOwnProperty("separator") ? configVar["separator"] : " ";
+				customInput = _openfiles.join(separator);
+				customInputWsl = _openfiles.map(path => "\"" + Util.getWSLPath(path) + "\"").join(separator);
+				customInputDocker = isDockerMode ? _openfiles.map(path => "\"" + Util.getDockerPath(path) + "\"").join(separator) : undefined;
 			}
 			
 			if(customInput){
 				if(this.options.isWslMode && Util.isWindows && customInputWsl){
-					if(task.winNativePath){
+					if(task.isNativeCommand){
 						output = Util.replaceAll(output, "\\" + matches[0], customInput);
 					} else {
+						XycodeUI.instance.debug("wsl path: " + customInputWsl);
 						output = Util.replaceAll(output, "\\" + matches[0], customInputWsl);
 					}
 					this.customVars["WSL__" + varkey] = customInputWsl;
+				} else if(matches[6] === '|wslpath'){
+					output = Util.replaceAll(output, "\\" + matches[0].replace('|wslpath', '\\|wslpath'), customInputWsl ? customInputWsl : Util.getWSLPath(customInput));
+				} else if(Util.isDockerMode && customInputDocker){
+					if(task.isNativeCommand){
+						output = Util.replaceAll(output, "\\" + matches[0], customInput);
+					} else {
+						XycodeUI.instance.debug("docker path: " + customInputDocker);
+						output = Util.replaceAll(output, "\\" + matches[0], customInputDocker);
+					}
 				} else {
 					output = Util.replaceAll(output, "\\" + matches[0], customInput);
 				}
 				this.customVars[varkey] = customInput;
 			} else {
-				throw new Error(`${task.command} \r\n ${varkey} : ${customInput} is null!`);
+				throw new Error(`${task.command} \n ${varkey} : ${customInput} is null!`);
 			}
 		}
-		if(this.options.isWslMode && Util.isWindows && !task.winNativePath){
-			output = this.replaceWslFileAttr(output, this.fileAttr);
+		if(!task.isNativeCommand){
+			if(this.options.isWslMode && Util.isWindows){
+				output = this.replacePosixFileAttr(output, this.fileAttr, "wsl");
+			}
+			if(Util.isDockerMode){
+				output = this.replacePosixFileAttr(output, this.fileAttr, "docker");
+			}
 		}
-		return this.format(output);
+		command = this.format(output);
+		command = this.formatAsPlatformCommand(task, command);
+		return command;
+	}
+
+	private init(file: string, workspaceFolder:string) {
+		this.fileAttr = this.getFileAttr(file, workspaceFolder);
+		this.options.dockerContainer = this.replaceFileAttr(Util.dockerContainer);
+	}
+
+	public formatAsPlatformCommand(task: TaskType, command: string): string{
+		XycodeUI.instance.info(command);
+		if(task.isNativeCommand) {
+			return command;
+		}
+		if(Util.isDockerMode){
+			if(!this.options.dockerContainer){
+				throw new Error(`can not find docker container!`);
+			}
+			command = task.dockerOptions?.cwd ? `cd "${task.dockerOptions?.cwd}" && ${command}` : command;
+			command = Util.replaceAll(command, "\"", "\\\"");
+			if(task.dockerOptions?.openTTY){
+				command = `docker exec -it ${this.options.dockerContainer} /bin/sh -c "${command}"`;
+			} else {
+				command = `docker exec ${this.options.dockerContainer} /bin/sh -c "${command}"`;
+			}
+			XycodeUI.instance.info(command);
+		}
+		if(Util.isWslMode && Util.isWindows && !task.termial){
+			// TODO
+			// task.termial.shellPath || this.options.shellPath
+			command = `"${this.options.shellPath}" -c "${Util.replaceAll(command, "\"", "\\\"")}"`;
+			XycodeUI.instance.info(command);
+		}
+		return command;
 	}
 
 	public format(variable: string | undefined): string {
+		XycodeUI.instance.debug("start to format command");
 		if(!variable) { return ""; }
 		let result = variable;
 		for(let key in this.customVars) {
+			if(!result.includes(key)) {
+				continue;
+			}
 			result = Util.replaceAll(result, "\\${" + key + "}", this.customVars[key]);
+			if(result.includes("|wslpath")) {
+				result = Util.replaceAll(result, "\\${" + key + "\\|wslpath}", this.customVars[key] ? Util.getWSLPath(this.customVars[key]) : "");
+			}
 		}
 		for(let key in this.configVars) {
+			if(!result.includes(key)) {
+				continue;
+			}
 			let val = this.configVars[key]["value"];
 			if(typeof val === "string" || typeof val === "number"){
 				result = Util.replaceAll(result, "\\${" + key + "}", val.toString());
+				if(result.includes("|wslpath")) {
+					result = Util.replaceAll(result, "\\${" + key + "\\|wslpath}", val.toString() ? Util.getWSLPath(val.toString()) : "");
+				}
 			}
 		}
 		return this.replaceFileAttr(result);
 	}
 
 	private replaceFileAttr(command:string): string {
+		XycodeUI.instance.debug("start to replaceFileAttr");
+		if(!command) {
+			return "";
+		}
 		for(let key in this.fileAttr) {
+			if(!command.includes(key)) {
+				continue;
+			}
 			command = Util.replaceAll(command, "\\${" + key + "}", this.fileAttr[key]); 
+			if(command.includes("|wslpath")) {
+				command = Util.replaceAll(command, "\\${" + key + "\\|wslpath}", this.fileAttr[key] ? Util.getWSLPath(this.fileAttr[key]) : "");
+			}
 		}
 		return command;
 	}
 
-	private replaceWslFileAttr(command:string, fileAttr: { [index: string]: string; }): string {
+	private replacePosixFileAttr(command:string,fileAttr: { [index: string]: string; }, type: string): string {
 		for(let key of ["HOME", "TMPDIR", "file", "fileDirname", "workspaceFolder"]) {
 			if(fileAttr.hasOwnProperty(key)){
-				command = Util.replaceAll(command, "\\${" + key + "}", Util.getWSLPath(fileAttr[key])); 
+				command = Util.replaceAll(command, "\\${" + key + "}", type === "wsl" ? Util.getWSLPath(fileAttr[key]) : type === "docker" ? Util.getDockerPath(fileAttr[key]) : fileAttr[key] ); 
 			}
 		}
 		return command;
@@ -217,13 +310,14 @@ export class CommandBuilder{
 			"fileBasenameNoExtension":  path.basename(file, path.extname(file)),
 			"relativeFile" : path.relative(workspaceFolder, file),
 			"relativeFileDirname" : path.relative(workspaceFolder, path.dirname(file)),
+            "lowercaseWorkspaceName" : Util.workspaceLowercaseName,
             "workspaceFolder" : workspaceFolder,
             "workspaceFolderBasename" : path.basename(workspaceFolder),
             "fileDirname" : path.dirname(file),
 			"fileExtname" : path.extname(file),
 			"selectedText" : Util.selectedText,
             "YYYYMMDD": YYYYMMDD,
-            "YYYYMMDD_HHmm": [YYYYMMDD, "_", HHmm].join("")
+			"YYYYMMDD_HHmm": [YYYYMMDD, "_", HHmm].join("")
 		};
 		return fileAttr;
     }
@@ -234,10 +328,18 @@ class DefaultEnv
 	constructor(private readonly options : CommandRunnerOptions){
 	}
 
-	public async getSfdxAlias() : Promise<{[index: string]: any;}> {
+	public async load(command:string) : Promise<{[index: string]: any;}>{
+		return new Promise<{[index: string]: any;}>(async (resolve, reject) => {
+			resolve({
+				"SFDX_ALIAS": command.includes("SFDX_ALIAS") ? await this.getSfdxAlias(this.options.isWslMode && Util.isWindows || Util.isDockerMode) : []
+			});
+		});
+	}
+
+	private async getSfdxAlias(isWsl:boolean) : Promise<{[index: string]: any;}> {
 		return {
 				"label" : "Sfdx Alias",
-				"value" : this.options.isWslMode && Util.isWindows ? await this.getWslSfdxAlias() : this._getSfdxAlias()
+				"value" : isWsl ? await this.getWslSfdxAlias() : this._getSfdxAlias()
 		};
 	}
 
@@ -258,15 +360,27 @@ class DefaultEnv
 	private async getWslSfdxAlias() : Promise<Array<SimpleQuickPickItem>>  {
 		return new Promise<Array<SimpleQuickPickItem>>(async (resolve, reject) => {
 			try{
-				const command = 'cat ~/.sfdx/alias.json';
+				let command = 'cat ~/.sfdx/alias.json';
 				let options :child_process.ExecOptions = { 
 					maxBuffer: this.options.maxBuffer, 
 					shell:  this.options.shellPath
 				};
+				if(Util.isDockerMode){
+					options.cwd = Util.workspaceFolder;
+					command = `docker exec ${this.options.dockerContainer} /bin/sh -c "${command}"`;
+				}
 				const { stdout, stderr } = await new BaseCommandRunner().exec(command, options);
+				if(stderr) {
+					XycodeUI.instance.debug(stderr);
+				}
+				if(stdout) {
+					XycodeUI.instance.debug("load alias.json done!");
+					XycodeUI.instance.debug(stdout);
+				}
 				const orgs = JSON.parse(stdout)['orgs'];
 				resolve(this.getSfdxAliasQuickPickItems(orgs));
-			}catch(e){
+			} catch(e){
+				XycodeUI.instance.warn("get SfdxAlias exception : " + e);
 				resolve([]);
 			}
 		});